18/4/20
    //过滤自己的订单 表加入masterUID 外键over
    //安全性校验 判断UID与订单是否匹配getuserInfo getdetaildeinfo dissent  over
    //订单UID生成函数修改数字 over

+++ app/dissent.php
@@ -10,6 +10,7 @@
     else{
         $info = new errorInfo("未登录");
         Tools::infoBack($info);
+        return;
     }
 
     function dissent(){
@@ -18,11 +19,20 @@
         $databaseTools = new databaseTools();
         $database = $databaseTools->databaseInit();
 
-        $result = $database->select('waiting_finish','form_id',array('form_id'=>$formID));
+        $result = $database->select('waiting_finish',array(
+            '[>]user_basic'=>array('form_id'=>'formID')
+        ),array('form_id','masterUID','serverUID'),array('form_id'=>$formID));
+
         if(empty($result)){
             $info = new errorInfo("订单未结算或不存在");
             Tools::infoBack($info);
         }else{
+            $result = $result[0];
+            if($result['masterUID']!=$_SESSION['UID']&&$_SESSION['UID']!=$result['serverUID']){
+                $info = new errorInfo("该用户没有权限");
+                Tools::infoBack($info);
+                return;
+            }
             $database->insert('waiting_judge',array('form_id'=>$formID,'bz'=>$bz));
             $database->update('form_basic',array('flag'=>4),array('formID'=>$formID));
             $database->delete('waiting_finish',array('form_id'=>$formID));
